+++ dart/lib/src/noop_hub.dart
@@ -75,7 +75,10 @@ class NoOpHub implements Hub {
   Future<void> addBreadcrumb(Breadcrumb crumb, {dynamic hint}) async {}
 
   @override
-  Future<SentryId> captureTransaction(SentryTransaction transaction) async =>
+  Future<SentryId> captureTransaction(
+    SentryTransaction transaction, {
+    SentryTraceContextHeader? traceContext,
+  }) async =>
       SentryId.empty();
 
   @override