+++ dart/lib/src/sentry_span_interface.dart
@@ -52,8 +52,16 @@ abstract class ISentrySpan {
   set throwable(dynamic throwable);
 
   @internal
-  bool? get sampled;
+  SentryTracesSamplingDecision? get samplingDecision;
 
   /// Returns the trace information that could be sent as a sentry-trace header.
   SentryTraceHeader toSentryTrace();
+
+  /// Returns the baggage that can be sent as "baggage" header.
+  @experimental
+  SentryBaggageHeader? toBaggageHeader();
+
+  /// Returns the trace context.
+  @experimental
+  SentryTraceContextHeader? traceContext();
 }